svms-job-module-service/requirements.txt
aiohttp==3.7.3
amqp==5.0.2
APScheduler==3.6.3
asgiref==3.3.0
async-timeout==3.0.1
asyncio==3.4.3
attrs==20.2.0
bcrypt==3.2.0
billiard==3.6.3.0
boto3==1.14.59
botocore==1.17.59
bson==0.5.8
cached-property==1.5.2
cachetools==4.1.0
celery==5.0.2
certifi==2020.4.5.1
cffi==1.14.3
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-repl==0.1.6
configparser==5.0.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.2.1
distro==1.5.0
Django==3.0.8
django-celery-beat==2.1.0
django-celery-results==2.0.0
django-cors-headers==3.2.1
django-extensions==3.0.9
django-filter==2.4.0
django-mysql==3.8.1
django-redis==4.12.1
django-simple-history==2.11.0
django-timezone-field==4.0
django-unixtimestampfield==0.4.0
django-yamlconf==1.3.0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
djangotoolbox==1.8.0
docker==4.3.1
docker-compose==1.27.4
dockerpty==0.4.1
docopt==0.6.2
docutils==0.15.2
drf-yasg==1.17.1
et-xmlfile==1.0.1
future==0.18.2
google-api-core==1.22.2
google-api-python-client==1.12.5
google-auth==1.21.3
google-auth-httplib2==0.0.4
googleapis-common-protos==1.52.0
googlemaps==4.4.2
grpcio==1.32.0
grpcio-tools==1.32.0
gunicorn==20.0.4
h11==0.11.0
hiredis==1.0.1
httplib2==0.18.1
httptools==0.1.1
hyperlink==19.0.0
idna==2.10
importlib-metadata==1.6.1
inflection==0.5.1
itypes==1.2.0
jdcal==1.4.1
Jinja2==2.11.2
jmespath==0.10.0
jsonfield==3.1.0
jsonpickle==1.4.2
jsonschema==3.2.0
kombu==5.0.2
lml==0.1.0
MarkupSafe==1.1.1
msgpack==0.6.2
multidict==5.1.0
mysqlclient==1.4.6
openpyxl==3.0.5
packaging==20.4
paramiko==2.7.2
Pillow==7.1.2
prompt-toolkit==3.0.8
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyexcel-io==0.6.4
pyexcel-xls==0.6.1
pyexcel-xlsx==0.6.0
PyHamcrest==2.0.2
PyJWT==1.7.1
pymongo==3.11.0
PyMySQL==0.10.1
PyNaCl==1.4.0
pyparsing==2.4.7
pyrsistent==0.17.3
PySocks==1.7.1
python-crontab==2.5.1
python-dateutil==2.8.1
python-decouple==3.3
python-dotenv==0.15.0
python-http-client==3.2.7
pytz==2019.3
PyYAML==5.3.1
redis==3.5.3
requests==2.23.0
requests-futures==1.0.0
rsa==4.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
s3transfer==0.3.3
sendgrid==6.2.2
service-identity==18.1.0
six==1.14.0
sqlparse==0.2.4
stomp.py==6.1.0
texttable==1.6.3
typing-extensions==3.7.4.3
tzlocal==2.1
uritemplate==3.0.1
urllib3==1.25.11
uvicorn==0.12.2
uvloop==0.14.0
vine==5.0.0
wcwidth==0.2.5
websocket-client==0.57.0
wrapt==1.12.1
xlrd==1.2.0
xlwt==1.3.0
yarl==1.6.3
zipp==3.4.0
aws-xray-sdk==2.6.0
health-check==3.4.1
django-admin-rangefilter==0.6.3
